requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
httpx>=0.28.0

# Database
supabase>=2.10.0

# AI/ML Models
torch>=2.5.0
transformers>=4.46.0
accelerate>=1.1.0
sentencepiece>=0.2.0
hf-transfer>=0.1.8

# LangChain & LangGraph
langgraph>=0.2.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain>=0.3.0

# NLP & Text Processing
rapidfuzz>=3.10.0

# Search & Web
tavily-python>=0.5.0
duckduckgo-search>=6.0.0

# UI
streamlit>=1.40.0

# HTTP Requests
requests>=2.32.0

# Testing
pytest>=8.3.0

# Optional: OpenAI for fallback (install if you have API key)
# openai>=1.0.0
